+++ docs/source/source/_bss_dm.rst
@@ -1,0 +1,8 @@
+.. _bss_dm:
+
+APS DM interface to APS Scheduling System
+=========================================
+
+.. automodule:: apsbss.bss_dm
+    :members:
+
