+++ docs/source/source/_core.rst
@@ -1,0 +1,8 @@
+.. _core:
+
+Common Support for apsbss package
+=================================
+
+.. automodule:: apsbss.core
+    :members:
+
